Why is my code not working
my on clicked event for a button in the main ui has a bit of code that
is identical to a bit of code in the on clicked event for a button in
the info resource image.  The code in the main ui works, but the other
code does not.